feat: Support [^reviewing] section tag.
This adds support for a special syntax to add a "reviewing" badge
after an <h3>

The syntax is:

    ### Header

    [^reviewing]

    Lorem ipsum dolor set amet...

This syntax was selected because it mirrors the Markdown footnote
syntax and is unlikely to conflict with anything. The actual
transformation is handled in JavaScript for now, rather than the
Markdown parser. The Markdown parser would be arguably _better_
but requires significantly more work. This may be refactored in
the future.