Update server so permanent links aren't broken
When rendered to an output the site generator provides redirects and directories for the scopes. The flask server version did NOT support this and cause all permalinks to be broken when running locally in a flask server.  This should remedy this and provide 404s instead of key errors in the flask app when a page doesn't exist.