chore: adding container tooling
Since site-generator is stuck on Python 3.8, this will make
it difficult to run the site-generator in a local environment.

Adding a Dockerfile and serve.sh similar to google.aip.dev to
allow easy local execution.

the gitignore was added because some python installations create
a __pycache__ directory.

setup.py's pyyaml was bumped to work around a backwards-incompatible
Cython upgrade to 3.0.